ruoyi-fastapi-backend/requirements.txt
APScheduler==3.11.0
asyncmy==0.2.10
DateTime==5.5
fastapi[all]==0.115.8
loguru==0.7.3
openpyxl==3.1.5
pandas==2.2.3
passlib[bcrypt]==1.7.4
Pillow==11.1.0
psutil==7.0.0
pydantic-validation-decorator==0.1.4
PyJWT[crypto]==2.10.1
PyMySQL==1.1.1
redis==5.2.1
requests==2.32.3
SQLAlchemy[asyncio]==2.0.38
sqlglot[rs]==26.6.0
user-agents==2.2.0
akshare
scikit-learn
scikit-optimize
pandas_market_calendars
openai
mcp
torch
statsmodels
tensorflow
matplotlib
seaborn
aiofiles